Wires up user sign-ins
Users can sign in with a password or an email. When a user enters their email address, if there is an associated password we show the "password" field.  Otherwise we just show the "sign in" field which sends a magic link.  This link is a tokenized URL which expires in 15 minutes.

There is a bit of leftover views from the Phoenix auth generator, I'll be cleaning those up as I go.

A bunch of other styling features included in this PR.  I think we are at a place where this can be deployed to Fly in the next PR.