Adds about component for places
Lots of small changes, fixes, etc.
The largest part of this was adding an `About` component that includes photo uploads for place avatars.

I noticed that sometimes images would crash the Fly instance (I need to upsize the instance) but in those cases the images wouldn't be resized, so I've added a new call to `new_resize_images/1` if there are no variants available... I want to be sure nobody is downloading meg + images on their smart phone.

Further image improvements coming soon.