Owners can update user roles in their places
This is very ugly at the moment and I'm not proud of it.  It's the fastest way to get this feature up and working.

Also a small fix for email sign in links, sometimes the hash would have an additional "-" which was preventing this from finding the token when the string split the remember me token.